requirements.txt
einops==0.6.1
diffusers==0.16.1
transformers==4.28.1
pytorch-lightning==1.5.0
torch==2.0.0
opencv-python==4.7.0.72
protobuf==3.20.3
wrapt==1.15.0
numpy==1.23.5
google==3.0.0
googleapis-common-protos==1.59.0

# Inference service (FastAPI + queue)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
redis>=5.0.0
pydantic-settings>=2.0.0